+++ CHANGELOG.md
@@ -18,6 +18,7 @@
   - Introduced `BooleanBag`
   - Introduced `CharBag`
 - Removed `Mapper.add(int, Component)`, `create` should be used instead
+- `Mask.set(Mask/long[])` renamed to `copyFrom`
 - Fix: `ByteBag` implements `ensureCapacity(int)`
 - Fix: `Bag.ensureCapacity(int)` throws `NegativeArraySizeException` if `capacity` is negative.